frontend/requirements.txt
streamlit
st_annotated_text
requests
